+++ backend/dissemination/views.py
@@ -12,8 +12,6 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.utils import timezone
 from django.views.generic import View
 
-from audit.decorators import newrelic_timing_metric
-
 from config.settings import STATE_ABBREVS, SUMMARY_REPORT_DOWNLOAD_LIMIT
 
 from dissemination.file_downloads import get_download_url, get_filename
@@ -33,6 +31,7 @@ from dissemination.models import (
     OneTimeAccess,
 )
 from dissemination.summary_reports import generate_summary_report
+from support.decorators import newrelic_timing_metric
 
 from users.permissions import can_read_tribal
